Overhaul thumbnail view
The thumbnail view is overloaded and confusing, especially for
multiple tracks. This patch implements the design proposed
in #1624 to help with that.

This is mostly just a major restyling, so no important
functionality should be lost.
